chat-main/requirements.txt
uvicorn~=0.34.0
fastapi~=0.115.10
openai~=1.65.1
firecrawl-py
gunicorn
pymongo~=4.11.1
dnspython~=2.7.0
certifi~=2025.1.31
langchain-openai~=0.3.7
langchain-milvus